+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added PropTypes for ariaLabel on Cell
 
+### Changed
+* Updated unmount implementation to cancel any pending animation frames
+
 2.2.0 - (February 5, 2019)
 ----------
 ### Added